backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-environ>=0.10,<1.0
psycopg2-binary>=2.9,<3.0
